#40 added log daemon

+++ apps/utils/cppUtils.cpp
@@ -2,12 +2,12 @@
 #include "cppUtils.hpp"
 #include <lib/syscall.h>
 
-void *operator new(long unsigned int size)
+void *operator new(unsigned int size)
 {
     return malloc(size);
 }
  
-void *operator new[](long unsigned int size)
+void *operator new[](unsigned int size)
 {
     return malloc(size);
 }
